services/user_service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
pydantic[email]==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
email-validator==1.1.3
bcrypt==3.2.0
PyJWT==2.1.0